sock: add ee_code SO_EE_CODE_ZEROCOPY_COPIED
A send request with MSG_ZEROCOPY falls back to copying when a request
cannot be completed in zerocopy mode.

Notify processes when the call reverted to copying through a new bit
in the completion notification, so that they can optionally decline
passing the flag on subsequent calls on the same path.

Signed-off-by: Willem de Bruijn <willemb@google.com>