select: compute slack based on relative time
Avoid one ktime_get_ts64 when computing hrtimer slack.

Select and poll pass the timeout as a timespec64 in absolute time.
Absolute time is needed both to program schedule_hrtimeout_range
and to compute the remainder on return to userspace.

But the current method converts from relative to absolute and
back to relative time when computing slack. Avoid the latter,
by computing slack before the conversion to absolute.

This saves one expensive clock read, but at the cost of having
to pass a u64 slack through a few functions.

Not-yet-signed-off-by: Willem de Bruijn <willemb@google.com>