tcp: pass device memory geometry iovec as cmsg
In tcp_recvmsg

1. detect when data is backed by p2pdma pages and if so
2. pass the GPU mem metadata as cmsg.

The host CPU does not have to have access to these device pages.

A tcp reader can use MSG_TRUNC to skip the copy to user, but still
receive the original data length as return value, plus optional cmsgs
with placement info in GPU memory.

In practice, pages may very well be mapped and accessible.

The current proof-of-concept returns two offsets:
a. to the start of the region, to pass to cuMemcpyDtoD
b. an opaque pointer to return to the kernel when the buffer
   is to be released.

(Second is adding in SO_DEVMEM_OFFSET follow-on patch. Obviously
 should be combined. Should not require two put_cmsgs either.)

This second buffer is currently a raw physical address. This is
obviously unsafe and should be some opaque handle checked on return to
kernel.