net: add SO_DEVMEM_OFFSET setsockopt to release pages
Add an interface to notify the kernel that it can recycle buffers.

When receiving data in p2pdma device pages, applications receive an
opaque buffer handle in cmsg SCM_DEVMEM_OFFSET.

Add setsockopt SO_DEVMEM_OFFSET to return these handles to the kernel,
to notify it that the pages are no longer in use and can be reused.

This is very much an unsafe proof-of-concept implementation that
relies on trusted applications to only return valid handles, once.